requirements.txt
# Jira Workflow Tools - Core Dependencies
# Updated for team version with private mode support

# Core HTTP and API client
requests>=2.31.0

# Command-line interface framework
click>=8.1.0

# Rich terminal formatting and display
rich>=13.0.0

# Environment variable management
python-dotenv>=1.0.0

# YAML parsing for configuration and templates
PyYAML>=6.0.0

# Template engine for customizable issue templates
Jinja2>=3.1.0

# Development and testing dependencies
# Install with: pip install -r requirements-dev.txt